refactor(crane-context): split memory-curator.ts; extract scoring + apply helpers
Brings workers/crane-context/src/lib/memory-curator.ts under structural caps.

- memory-curator.ts (308 LOC, was 445): orchestrator + scoreContradiction
  reduced to 31 lines, runMemoryCurator to 16 lines.
- memory-curator-helpers.ts: buildFtsExpr, queryTopFtsMatch,
  invokeAiContradiction, parseAiFirstLine, extractAiRaw, applyNoteUpdates
  (split into applyPromotions + persistScoreRow to satisfy 75-line cap).

scoreContradiction complexity 20 → <10. runMemoryCurator complexity 18 → <10.
no-useless-assignment on topMatch fixed via direct assignment.
Public API preserved. 540/540 tests pass.
